Use ListObjectsPages instead of "buggy" ListObjectsV2Pages
The use of *V2* API is recommended in the docs, however, I
can prove that the ListObjectsV2Pages was returning only
555 Prefixes instead of 786.

The ListObjectsPages and s3cmd cli is implemented using
`Marker` in the prefix list instead of the continuation
token (ListObjectsV2Pages) implementation. The former
has proven to give consistent results with this bucket and
the later is source of bugs for us.